requirements.txt
absl-py
astunparse
cached-property
cachetools
certifi
charset-normalizer
clang
click
colorama
cycler
Cython
decorator
Flask
flatbuffers
fvcore
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
gunicorn
h5py
idna
imageio
imgviz
importlib-metadata
itsdangerous
Jinja2
keras
Keras-Preprocessing
kiwisolver
Markdown
MarkupSafe
matplotlib
natsort
networkx
ninja
numpy
oauthlib
opencv-python
opt-einsum
packaging
Pillow
protobuf
pyasn1
pyasn1-modules
pycocotools
pyparsing
python-dateutil
PyWavelets
PyYAML
QtPy
requests
requests-oauthlib
rsa
scikit-image
scipy
setuptools
six
tensorflow
tensorflow-estimator
termcolor
torch
torchvision
tifffile
typing-extensions
urllib3
Werkzeug
wrapt
zipp
